move LabelMaker to /lib/charms/
