+++ tox.ini
@@ -9,7 +9,7 @@ envlist = lint,unit
 
 [testenv]
 setenv =
-    PYTHONPATH={toxinidir}
+    PYTHONPATH={toxinidir}:{toxinidir}/lib
     PYTHONBREAKPOINT=ipdb.set_trace
 passenv = OS_* TEST_*
 
@@ -26,5 +26,5 @@ deps =
     black
     flake8
 commands =
-    flake8 {toxinidir}/reactive {toxinidir}/tests
-    black --check {toxinidir}/reactive {toxinidir}/tests
+    flake8 {toxinidir}/lib {toxinidir}/reactive {toxinidir}/tests
+    black --check {toxinidir}/lib {toxinidir}/reactive {toxinidir}/tests
